moving FdbPodClient from internal to pkg
